chore: update applyDiscount to use BigDecimal for discount value

+++ chapter-02/src/main/java/expert/os/books/architecture/patterns/scenario01/rich/Customer.java
@@ -1,12 +1,14 @@
 package expert.os.books.architecture.patterns.scenario01.rich;
 
+import java.math.BigDecimal;
+
 public class Customer {
     private int loyaltyPoints;
 
     // WE TELL: Business logic is co-located with the state it modifies.
     public void processCheckoutRewards(Order order) {
         if (this.loyaltyPoints > 100) {
-            order.applyDiscount(10.0);
+            order.applyDiscount(BigDecimal.valueOf(10.0));
             this.loyaltyPoints -= 100;
         }
         this.loyaltyPoints += 10;
